chore: Update Angular dependencies to version 17.3.11
This commit updates the Angular dependencies in the `package.json` file to version 17.3.11. It also updates the devDependencies for Angular CLI and compiler-cli to the same version. Additionally, the `zone.js` dependency is updated to version 0.14.7. The `angular.json` file is modified to include the `@angular/material/prebuilt-themes/indigo-pink.css` style.